# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Course Registration API",
        "version": "1.0.0",
        "title": "Course API",
        "contact": {
            "email": "atilola.oyediji@gmail.com"
        }
       
    },
    "schemes": ["http"],
    "host": "localhost:4000",
    "basePath": "",
    "paths" : {
        "/departments" : {
            "get" : {
                "summary" : "Get all departments in an institution",
                "description": "Get all courses",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/department"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post" : {
                "summary" : "Add New Department",
                "description": "Save a new department",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "course registration data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/department"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                                "$ref": "#/definitions/department"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "put" : {
                "summary" : "update department information",
                "description": "change name or faculty a department belongs",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "student data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/department"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                                "$ref": "#/definitions/department"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },

        "/students" : {
            "get" : {
                "summary" : "Get all the students",
                "description": "Get all the courses",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/student"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post" : {
                "summary" : "Add New Student",
                "description": "Save a new student",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "course registration data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/student"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                                "$ref": "#/definitions/student"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "put" : {
                "summary" : "Update Level of individual student e.g. 200 to 300",
                "description": "change student level",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "student data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/student"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                                "$ref": "#/definitions/student"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },

        "/students/{matricno}" : {        
            "get" : {
                "summary" : "get a student basic information by matriculation number",
                "description": "get single student information",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "matricno",
                        "in": "path",
                        "description": "matric number (unique) for individual student",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/student"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },

"/students/{id}" : {        
            "get" : {
                "summary" : "get a student basic information by id",
                "description": "get single student information",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "unique objectid",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/student"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },

        "/courses" : {
            "get" : {
                "summary" : "Get all courses",
                "description": "Get all the courses",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/course"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post" : {
                "summary" : "Add New Course",
                "description": "Save a new course",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "course registration data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/course"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                                "$ref": "#/definitions/course"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },


        "/registrations" : {
            "get" : {
                "summary" : "Get all registrations",
                "description": "Get all registered students",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/registration"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post" : {
                "summary" : "Add New Registration per semester",
                "description": "register a new course list",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "course registration data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/registration"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                                "$ref": "#/definitions/registration"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },

        "/registrations/student" : {
            "post" : {
                "summary" : "Get Student's Registration Based on Matric Number and Session",
                "description": "Get a New Registration",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "registration retrieved data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/registrations"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                                "$ref": "#/definitions/registration"
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }    
    }, 
    "definitions": {
        "department": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Political Science"
               },
               "faculty": {
                   "type": "string",
                   "example": "Sciences"
               },
               "isactive": {
                   "type": "boolean",
                   "example": true
               }

                }
        },
        "student": {
            "type": "object",
            "properties": {
                "surname": {
                    "type": "string",
                    "example": "Oyediji"
               },
               "firstname": {
                   "type": "string",
                   "example": "Atilola"
               },
               "middlename": {
                   "type": "string",
                   "example": "Joshua"
               },
               "department": {
                   "type": "string",
                   "example": "Computer Engineering"
               },
               "matricno": {
                "type": "string",
                "example": "CSC/2011/121"
                },
                "gender": {
                        "type": "string",
                        "example": "Male"
                        },
                "level": {
                        "type": "integer",
                        "example": "200"
                        },
                "yearofentry": {
                        "type": "integer",
                        "example": "2012"
                        }
                }
        },
        "course": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string",
                    "example": "Software Engineering"
               },
               "code": {
                   "type": "string",
                   "example": "SEE 112"
               },
               "departments": {
                   "type": "array",
                   "example": ["Computer Engineering", "Electrical Engineering", "Computer Science and Mathematics"]
               },
               "level": {
                "type": "integer",
                "example": "100"
                },
                "catcap": {
                "type": "integer",
                "example": "30"
                },
                "examcap": {
                "type": "integer",
                "example": "70"
                },
                "semester": {
                        "type": "string",
                        "example": "harmattan"
                },
                "units": {
                        "type": "integer",
                        "example": "2"
                },
                "type": {
                        "type": "string",
                        "example": "elective"
                }
            }
        },
        "registration": {
            "type": "object",
            "properties": {
                "registrationNumber": {
                    "type": "string",
                    "example": "RH2122"
               },
               "registrant": {
                   "type": "string",
                   "example": "6f9e30b47084d971a7c820ffb"
               },
               "session": {
                   "type": "string",
                   "example": 2012
               },
               "semester": {
                "type": "string",
                "example": "Rain"
                },
                "courses": {
                        "type": "array",
                        "example": ["6dhgsgsggs", "436sdgygsygysdgygdy"]
                        }
                }
        },
       "registrations": {
            "type": "object",
            "properties": {
               "registrant": {
                   "type": "string",
                   "example": "6f9e30b47084d971a7"
               },
            "session": {
                   "type": "string",
                   "example": "2012/2013"
               }

               
                }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "errors":{
                   "type": "array",
                   "items":{
                        "$ref": "#/definitions/error"
                   }
                } 
            }

        },
        "error":{
            "type":"object",
            "properties": {
                "message": {
                    "type":"string",
                    "example":"Error message"
                },
                "field": {
                    "type":"string",
                    "example":"error field if any"
                }
            }  
        }

        }
       
    }